+++ src/Indice.Features.Cases.AspNetCore/Controllers/AdminNotificationsController.cs
@@ -1,5 +1,6 @@
 ﻿using Indice.Features.Cases.Interfaces;
 using Indice.Features.Cases.Models.Responses;
+using Indice.Types;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -16,24 +17,31 @@ namespace Indice.Features.Cases.Controllers
     [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(ProblemDetails))]
     [ProducesResponseType(StatusCodes.Status403Forbidden, Type = typeof(ProblemDetails))]
     [ProducesResponseType(StatusCodes.Status500InternalServerError, Type = typeof(ProblemDetails))]
-    [Route("[casesApiPrefix]/manage/users")]
-    internal class AdminUsersController : ControllerBase
+    [Route("[casesApiPrefix]/manage/my/notifications")]
+    internal class AdminNotificationsController : ControllerBase
     {
-        private readonly ICaseTypeNotificationSubscriptionService _service;
+        private readonly INotificationSubscriptionService _service;
+        private readonly CasesApiOptions _casesApiOptions;
 
-        public AdminUsersController(ICaseTypeNotificationSubscriptionService service) {
+        public AdminNotificationsController(
+            INotificationSubscriptionService service,
+            CasesApiOptions casesApiOptions) {
             _service = service ?? throw new ArgumentNullException(nameof(service));
+            _casesApiOptions = casesApiOptions ?? throw new ArgumentNullException(nameof(casesApiOptions));
         }
 
         /// <summary>
         /// Get the case type subscriptions of a user.
         /// </summary>
         /// <returns></returns>
-        [HttpGet("subscriptions")]
-        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(CaseTypeSubscription))]
-        public async Task<IActionResult> GetCaseTypeNotificationSubscription() {
-            var subscribed = await _service.GetCaseTypeNotificationSubscriptionByUser(User);
-            return Ok(new CaseTypeSubscription {
+        [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(NotificationSubscriptionResult))]
+        public async Task<IActionResult> GetMySubscriptions() {
+            var options = new ListOptions<NotificationFilter> {
+                Filter = NotificationFilter.FromUser(User, _casesApiOptions.GroupIdClaimType)
+            };
+            var subscribed = await _service.GetSubscriptions(options);
+            return Ok(new NotificationSubscriptionResult {
                 Subscribed = subscribed
             });
         }
@@ -42,10 +50,10 @@ namespace Indice.Features.Cases.Controllers
         /// Create new case type subscription for a user.
         /// </summary>
         /// <returns></returns>
-        [HttpPost("subscriptions")]
+        [HttpPost]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
-        public async Task<IActionResult> CreateCaseTypeNotificationSubscription() {
-            await _service.CreateCaseTypeNotificationSubscription(User);
+        public async Task<IActionResult> Subscribe() {
+            await _service.Subscribe(NotificationSubscription.FromUser(User, _casesApiOptions.GroupIdClaimType));
             return NoContent();
         }
 
@@ -53,10 +61,10 @@ namespace Indice.Features.Cases.Controllers
         /// Remove a case type subscription for a user.
         /// </summary>
         /// <returns></returns>
-        [HttpDelete("subscriptions")]
+        [HttpDelete]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
-        public async Task<IActionResult> DeleteCaseTypeNotificationSubscription() {
-            await _service.DeleteCaseTypeNotificationSubscriptionByUser(User);
+        public async Task<IActionResult> Unsubscribe() {
+            await _service.Unsubscribe(NotificationFilter.FromUser(User, _casesApiOptions.GroupIdClaimType));
             return NoContent();
         }
     }