+++ src/Indice.Features.Cases.AspNetCore/Data/Config/DbCaseConfiguration.cs
@@ -1,5 +1,6 @@
 ﻿using Indice.Configuration;
 using Indice.Features.Cases.Data.Models;
+using Indice.Features.Cases.Extensions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
@@ -13,26 +14,10 @@ namespace Indice.Features.Cases.Data.Config
             builder
                 .HasKey(p => p.Id);
             builder
-                .OwnsOne(
-                    p => p.CreatedBy,
-                    actionBuilder => {
-                        var prefix = nameof(DbCase.CreatedBy);
-                        actionBuilder
-                            .Property(p => p.Id)
-                            .HasColumnName($"{prefix}{nameof(DbCase.CreatedBy.Id)}")
-                            .HasMaxLength(TextSizePresets.S64);
-                        actionBuilder
-                            .Property(p => p.Email)
-                            .HasColumnName($"{prefix}{nameof(DbCase.CreatedBy.Email)}")
-                            .HasMaxLength(TextSizePresets.M128);
-                        actionBuilder
-                            .Property(p => p.Name)
-                            .HasColumnName($"{prefix}{nameof(DbCase.CreatedBy.Name)}")
-                            .HasMaxLength(TextSizePresets.M128);
-                        actionBuilder
-                            .Property(p => p.When)
-                            .HasColumnName($"{prefix}{nameof(DbCase.CreatedBy.When)}");
-                    })
+                .OwnsOneAudit(p => p.CreatedBy)
+                .OwnsOneAudit(p => p.CompletedBy)
+                .OwnsOneAudit(p => p.AssignedTo);
+            builder
                 .OwnsOne(
                     p => p.Customer,
                     actionBuilder => {
@@ -57,50 +42,6 @@ namespace Indice.Features.Cases.Data.Config
                             .Ignore(p => p.FullName);
                     });
             builder
-                .OwnsOne(
-                    p => p.CompletedBy,
-                    actionBuilder => {
-                        var prefix = nameof(DbCase.CompletedBy);
-                        actionBuilder
-                            .Property(p => p.Id)
-                            .HasColumnName($"{prefix}{nameof(DbCase.CompletedBy.Id)}")
-                            .HasMaxLength(TextSizePresets.S64);
-                        actionBuilder
-                            .Property(p => p.Email)
-                            .HasColumnName($"{prefix}{nameof(DbCase.CompletedBy.Email)}")
-                            .HasMaxLength(TextSizePresets.M128);
-                        actionBuilder
-                            .Property(p => p.Name)
-                            .HasColumnName($"{prefix}{nameof(DbCase.CompletedBy.Name)}")
-                            .HasMaxLength(TextSizePresets.M128);
-                        actionBuilder
-                            .Property(p => p.When)
-                            .HasColumnName($"{prefix}{nameof(DbCase.CompletedBy.When)}");
-                    }
-                );
-            builder
-                .OwnsOne(
-                    p => p.AssignedTo,
-                    actionBuilder => {
-                        var prefix = nameof(DbCase.AssignedTo);
-                        actionBuilder
-                            .Property(p => p.Id)
-                            .HasColumnName($"{prefix}{nameof(DbCase.AssignedTo.Id)}")
-                            .HasMaxLength(TextSizePresets.S64);
-                        actionBuilder
-                            .Property(p => p.Email)
-                            .HasColumnName($"{prefix}{nameof(DbCase.AssignedTo.Email)}")
-                            .HasMaxLength(TextSizePresets.M128);
-                        actionBuilder
-                            .Property(p => p.Name)
-                            .HasColumnName($"{prefix}{nameof(DbCase.AssignedTo.Name)}")
-                            .HasMaxLength(TextSizePresets.M128);
-                        actionBuilder
-                            .Property(p => p.When)
-                            .HasColumnName($"{prefix}{nameof(DbCase.AssignedTo.When)}");
-                    }
-                );
-            builder
                 .Property(p => p.GroupId)
                 .HasMaxLength(TextSizePresets.M128);
             builder
@@ -118,12 +59,26 @@ namespace Indice.Features.Cases.Data.Config
                 .HasForeignKey(p => p.CheckpointId)
                 .OnDelete(DeleteBehavior.NoAction);
             builder
+                .HasOne(p => p.PublicData)
+                .WithMany()
+                .HasForeignKey(p => p.PublicDataId)
+                .OnDelete(DeleteBehavior.NoAction);
+            builder
+                .HasOne(p => p.Data)
+                .WithMany()
+                .HasForeignKey(p => p.DataId)
+                .OnDelete(DeleteBehavior.NoAction);
+            builder
                .Property(c => c.Metadata)
                .HasJsonConversion();
             builder
                 .Property(c => c.Channel)
                 .IsRequired()
                 .HasMaxLength(TextSizePresets.M128);
+            builder
+                .HasMany(p => p.Versions)
+                .WithOne(p => p.Case)
+                .HasForeignKey(p => p.CaseId);
         }
     }
 }