+++ src/Indice.Features.Cases.AspNetCore/Data/Config/DbCaseDataConfiguration.cs
@@ -1,5 +1,7 @@
-﻿using Indice.Configuration;
+﻿using System.Text.Json;
 using Indice.Features.Cases.Data.Models;
+using Indice.Features.Cases.Extensions;
+using Indice.Serialization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
@@ -7,32 +9,18 @@ namespace Indice.Features.Cases.Data.Config
 {
     internal class DbCaseDataConfiguration : IEntityTypeConfiguration<DbCaseData>
     {
+        public static readonly JsonSerializerOptions SerializerOptions = JsonSerializerOptionDefaults.GetDefaultSettings();
+
         public void Configure(EntityTypeBuilder<DbCaseData> builder) {
             builder
                 .ToTable("CaseData");
             builder
                 .HasKey(p => p.Id);
             builder
-                .OwnsOne(
-                    p => p.CreatedBy,
-                    actionBuilder => {
-                        var prefix = nameof(DbCaseData.CreatedBy);
-                        actionBuilder
-                            .Property(p => p.Id)
-                            .HasColumnName($"{prefix}{nameof(DbCaseData.CreatedBy.Id)}")
-                            .HasMaxLength(TextSizePresets.S64);
-                        actionBuilder
-                            .Property(p => p.Email)
-                            .HasColumnName($"{prefix}{nameof(DbCaseData.CreatedBy.Email)}")
-                            .HasMaxLength(TextSizePresets.M128);
-                        actionBuilder
-                            .Property(p => p.Name)
-                            .HasColumnName($"{prefix}{nameof(DbCaseData.CreatedBy.Name)}")
-                            .HasMaxLength(TextSizePresets.M128);
-                        actionBuilder
-                            .Property(p => p.When)
-                            .HasColumnName($"{prefix}{nameof(DbCaseData.CreatedBy.When)}");
-                    });
+                .OwnsOneAudit(p => p.CreatedBy);
+            builder
+                .Property(c => c.Data)
+                .HasJsonConversion();
         }
     }
 }